requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
<<<<<<< HEAD
Django==2.2.1
=======
Django==2.2.2
>>>>>>> 877893f7a98f908270bb5adb4f51ca251064d96c
django-cors-headers==3.0.2
django-cors-middleware==1.3.1
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
Fabric3==1.14.post1
gunicorn==19.9.0
idna==2.8
invoke==1.2.0
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
paramiko==2.4.2
psycopg2==2.7.1
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.2
